Load and store options to chrome.storage.sync, so that the last extension status is always up to date.
This status includes which feeds were already read, so that items in each feed that were not read
are highlighted when the popup is opened. The feed title in the popup is highlighted if it
contains unread items as well.
Once an item is opened (clicked) the "unread" status is taken off and the highlights are removed
appropriately.

Minor changes:
- Take all methods out of the objects to avoid store them in chrome.storage (maybe useless).
- add an utility object to compare dates, credits in the comments (utility.js file)