Reflect the order of feeds in settings permanent.
In the settings page you can now reorder the feed in the list by drag&drop operation
and the new order will be shown in the popup window.
To achieve this the options.Feeds array is changed and stored in the cloud at each
drag&drop operation. While this is a bit costly, it ensures that the preferences will always
be up to date (given the internet connection works).